+++ src/api/cart.ts
@@ -1,4 +1,4 @@
-import { apiStatus, apiError } from '../lib/util';
+import { apiStatus, apiError, getToken } from '../lib/util';
 import { Router } from 'express';
 import PlatformFactory from '../platform/factory';
 
@@ -13,11 +13,12 @@ export default ({ config, db }) => {
 
   /**
    * POST create a cart
-   * req.query.token - user token
+   * req.query.token | req.headers.authorization - user token
    */
   cartApi.post('/create', (req, res) => {
     const cartProxy = _getProxy(req)
-    cartProxy.create(req.query.token).then((result) => {
+    const token = getToken(req)
+    cartProxy.create(token).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -26,7 +27,7 @@ export default ({ config, db }) => {
 
   /**
    * POST update or add the cart item
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   body.cartItem: {
    *    sku: orderItem.sku,
    *    qty: orderItem.qty,
@@ -34,10 +35,11 @@ export default ({ config, db }) => {
    */
   cartApi.post('/update', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     if (!req.body.cartItem) {
       return apiStatus(res, 'No cartItem element provided within the request body', 500)
     }
-    cartProxy.update(req.query.token, req.query.cartId ? req.query.cartId : null, req.body.cartItem).then((result) => {
+    cartProxy.update(token, req.query.cartId ? req.query.cartId : null, req.body.cartItem).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -46,16 +48,17 @@ export default ({ config, db }) => {
 
   /**
    * POST apply the coupon code
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart Ids
    *   req.query.coupon - coupon
    */
   cartApi.post('/apply-coupon', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     if (!req.query.coupon) {
       return apiStatus(res, 'No coupon code provided', 500)
     }
-    cartProxy.applyCoupon(req.query.token, req.query.cartId ? req.query.cartId : null, req.query.coupon).then((result) => {
+    cartProxy.applyCoupon(token, req.query.cartId ? req.query.cartId : null, req.query.coupon).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -64,12 +67,13 @@ export default ({ config, db }) => {
 
   /**
    * POST remove the coupon code
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart Ids
    */
   cartApi.post('/delete-coupon', (req, res) => {
     const cartProxy = _getProxy(req)
-    cartProxy.deleteCoupon(req.query.token, req.query.cartId ? req.query.cartId : null).then((result) => {
+    const token = getToken(req)
+    cartProxy.deleteCoupon(token, req.query.cartId ? req.query.cartId : null).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -78,12 +82,13 @@ export default ({ config, db }) => {
 
   /**
    * GET get the applied coupon code
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart Ids
    */
   cartApi.get('/coupon', (req, res) => {
     const cartProxy = _getProxy(req)
-    cartProxy.getCoupon(req.query.token, req.query.cartId ? req.query.cartId : null).then((result) => {
+    const token = getToken(req)
+    cartProxy.getCoupon(token, req.query.cartId ? req.query.cartId : null).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -92,7 +97,7 @@ export default ({ config, db }) => {
 
   /**
    * POST delete the cart item
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   body.cartItem: {
    *    sku: orderItem.sku,
    *    qty: orderItem.qty,
@@ -100,10 +105,11 @@ export default ({ config, db }) => {
    */
   cartApi.post('/delete', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     if (!req.body.cartItem) {
       return apiStatus(res, 'No cartItem element provided within the request body', 500)
     }
-    cartProxy.delete(req.query.token, req.query.cartId ? req.query.cartId : null, req.body.cartItem).then((result) => {
+    cartProxy.delete(token, req.query.cartId ? req.query.cartId : null, req.body.cartItem).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -112,13 +118,14 @@ export default ({ config, db }) => {
 
   /**
    * GET pull the whole cart as it's currently se server side
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cartId
    */
   cartApi.get('/pull', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     res.setHeader('Cache-Control', 'no-cache, no-store');
-    cartProxy.pull(req.query.token, req.query.cartId ? req.query.cartId : null, req.body).then((result) => {
+    cartProxy.pull(token, req.query.cartId ? req.query.cartId : null, req.body).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -127,13 +134,14 @@ export default ({ config, db }) => {
 
   /**
    * GET totals the cart totals
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cartId
    */
   cartApi.get('/totals', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     res.setHeader('Cache-Control', 'no-cache, no-store');
-    cartProxy.totals(req.query.token, req.query.cartId ? req.query.cartId : null, req.body).then((result) => {
+    cartProxy.totals(token, req.query.cartId ? req.query.cartId : null, req.body).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -142,17 +150,18 @@ export default ({ config, db }) => {
 
   /**
    * POST /shipping-methods - available shipping methods for a given address
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart ID if user is logged in, cart token if not
    *   req.body.address - shipping address object
    */
   cartApi.post('/shipping-methods', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     res.setHeader('Cache-Control', 'no-cache, no-store');
     if (!req.body.address) {
       return apiStatus(res, 'No address element provided within the request body', 500)
     }
-    cartProxy.getShippingMethods(req.query.token, req.query.cartId ? req.query.cartId : null, req.body.address).then((result) => {
+    cartProxy.getShippingMethods(token, req.query.cartId ? req.query.cartId : null, req.body.address).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -161,13 +170,14 @@ export default ({ config, db }) => {
 
   /**
    * GET /payment-methods - available payment methods
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart ID if user is logged in, cart token if not
    */
   cartApi.get('/payment-methods', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     res.setHeader('Cache-Control', 'no-cache, no-store');
-    cartProxy.getPaymentMethods(req.query.token, req.query.cartId ? req.query.cartId : null).then((result) => {
+    cartProxy.getPaymentMethods(token, req.query.cartId ? req.query.cartId : null).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -176,17 +186,18 @@ export default ({ config, db }) => {
 
   /**
    * POST /shipping-information - set shipping information for collecting cart totals after address changed
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart ID if user is logged in, cart token if not
    *   req.body.addressInformation - shipping address object
    */
   cartApi.post('/shipping-information', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     res.setHeader('Cache-Control', 'no-cache, no-store');
     if (!req.body.addressInformation) {
       return apiStatus(res, 'No address element provided within the request body', 500)
     }
-    cartProxy.setShippingInformation(req.query.token, req.query.cartId ? req.query.cartId : null, req.body).then((result) => {
+    cartProxy.setShippingInformation(token, req.query.cartId ? req.query.cartId : null, req.body).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);
@@ -195,17 +206,18 @@ export default ({ config, db }) => {
 
   /**
    * POST /collect-totals - collect cart totals after shipping address changed
-   *   req.query.token - user token
+   *   req.query.token | req.headers.authorization - user token
    *   req.query.cartId - cart ID if user is logged in, cart token if not
    *   req.body.shippingMethod - shipping and payment methods object
    */
   cartApi.post('/collect-totals', (req, res) => {
     const cartProxy = _getProxy(req)
+    const token = getToken(req)
     res.setHeader('Cache-Control', 'no-cache, no-store');
     if (!req.body.methods) {
       return apiStatus(res, 'No shipping and payment methods element provided within the request body', 500)
     }
-    cartProxy.collectTotals(req.query.token, req.query.cartId ? req.query.cartId : null, req.body.methods).then((result) => {
+    cartProxy.collectTotals(token, req.query.cartId ? req.query.cartId : null, req.body.methods).then((result) => {
       apiStatus(res, result, 200);
     }).catch(err => {
       apiError(res, err);